say: Fix logging in log_rotate
log_rotate writes informational message only in plain format,
which is inappropriate when logger is configured to log in json format.

Fix it with replacing write with say_info which is safe
because it is ev signal callback not signal handler.

Fix bug say_format in json formatting,
Log message was invalid in several cases.

Closes #2987