sql: release all region memory after VDBE's halt
When ticket #3021 was implemented, alongside with it leak of region
memory was introduced. It occured due to allocation memory for tuples at
region at once (previously, it was allocated using malloc and before
insertion to space it was copied to the region).  However, if space is
ephemeral, it has nothing in common with txn routine. As a result,
region memory for tuples to be inserted in ephemeral spaces isn't pinned
to any txn and isn't freed by txn_commit() or txn_rollback(). To avoid
this leak, now  all used region memory is released in vdbe's halt
routine.

Closes #3199